LayoutTests/fast/events/spatial-navigation/snav-clipped-overflowed-content-expected.txt

That is is a normal link_1.

This is a scrollable Div created with the CSS property overflow.

It has a visible link_2.







... and an overflowed link like link_3.

And this is another normal link_4.

PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
